strip 1 component
