--- AutomotiveCluster/cpp/sources/main.cpp
@@ -1,24 +1,0 @@
-#include <QGuiApplication>
-#include <QQmlApplicationEngine>
-#include "circularindicator.h"
-
-int main(int argc, char *argv[])
-{
-#if QT_VERSION < QT_VERSION_CHECK(6, 0, 0)
-    QCoreApplication::setAttribute(Qt::AA_EnableHighDpiScaling);
-#endif
-
-    QGuiApplication app(argc, argv);
-
-    QQmlApplicationEngine engine;
-    qmlRegisterType<CircularIndicator>("ClusterDemo", 1, 0, "CircularIndicator");
-    const QUrl url(QStringLiteral("qrc:/main.qml"));
-    QObject::connect(&engine, &QQmlApplicationEngine::objectCreated,
-                     &app, [url](QObject *obj, const QUrl &objUrl) {
-        if (!obj && url == objUrl)
-            QCoreApplication::exit(-1);
-    }, Qt::QueuedConnection);
-    engine.load(url);
-
-    return app.exec();
-}
