--- AutomotiveCluster/qml/gui/GlobalDatas.qml
@@ -1,9 +1,0 @@
-import QtQuick 2.0
-
-Item {
-    property string perCharge: "0%"
-    property string kmCharge: "0"
-    property string imgCharge: "0"
-    property int countCharge: 0
-
-}
