+++ qml/gui.qrc
