+++ qml/gui/Motorama.qml
